requirements.txt
diffusers==0.32.2
pillow==11.1.0
transformers[sentencepiece]==4.48.1
config==0.5.1
opencv-python==4.11.0
einops==0.8.0
matplotlib==3.10.0
scipy==1.15.1
scikit-image==0.25.0
accelerate==1.3.0
timm==1.0.14
kornia==0.8.0